Finalized how we are going to authenticate
We will now authenticate on the client side. However after logging in we will save the id token generated after login into a cookie that is passed into the server end every time.
The server will validate this token to check if the user is authenticated or not properly

This makes it not on the server to be logged in but on the client and is properly handled where each client has to login to authenticate.

We should also be finally done with this nonsense and hence
Closes #2

Took 49 minutes