Implemented adding to database on confirmation of addition of data
If success or error, it will both redirect back to the add stats page. If it is a success message, there will be a success message. If the error is a duplicate entry (because a previous stats already exist), it will inform. Else there are no notifications

Took 1 hour 10 minutes