Making use of Multi-Arch docker builds
Docker builds will now be multi-arch on the master and tagged commits. Pulling the latest tag from the container will now provide you with the relevant image for your architecture. This applies to the release tags as well

Also release tags will now be found in its own dedicated Releases repository for sanity sake. Update your image ref accordingly. We will retain the old releases in the original state as they are not multi-arch

We also refactored our Dockerfile to include a build container image as apparently our Dockerfile was flawed lol. This caused our original build time to increase by 2 minutes as it is finally actually building the dependencies they should've built (previously not due to a lack of Python)